+++ client/src/components/Stores/Pickup/Pickup.vue
@@ -88,7 +88,7 @@
       :title="$i18n('pickup.join_title_date', $dateFormatter.dateTime(date))"
       :cancel-title="$i18n('pickup.join_cancel')"
       :ok-title="$i18n('pickup.join_agree')"
-      :ok-disabled="!loadedUserPickups || !pickupRulePass"
+      :ok-disabled="!loadedUserPickups || !pickupRulePass || isMissingHygieneCertificate"
       :ok-variant="okVariant"
       :hide-header-close="true"
       modal-class="bootstrap"
@@ -96,6 +96,13 @@
       lazy
       @ok="$emit('join', date)"
     >
+      <b-alert :show="isMissingHygieneCertificate" variant="danger">
+        <i class="fas fa-hands-wash" />
+        <span v-if="hygieneCertificateUntil" v-text="$i18n('pickup.hygieneCertificateMissing.timeout')" />
+        <span v-else v-text="$i18n('pickup.hygieneCertificateMissing.none')" />
+        <a :href="$url('settingsHygiene')" v-text="$i18n('pickup.hygieneCertificateMissing.link')" />
+      </b-alert>
+
       <p>{{ $i18n('pickup.really_join_date', slotInfo) }}</p>
 
       <div v-if="loadedUserPickups && sameDayPickups && sameDayPickups.length">
@@ -248,6 +255,8 @@ import { listSameDayPickupsForUser, checkPickupRuleStore } from '@/api/pickups'
 import TakenSlot from '@/components/Stores/Pickup/TakenSlot.vue'
 import EmptySlot from '@/components/Stores/Pickup/EmptySlot.vue'
 
+import StoreData from '@/stores/stores'
+
 export default {
   components: { EmptySlot, TakenSlot, BFormTextarea, BModal },
   directives: { VBTooltip },
@@ -316,6 +325,18 @@ export default {
     },
     emptySlots () {
       return Math.max(this.totalSlots - this.occupiedSlots.length, 0)
+    },
+    isHygieneRequired () {
+      return StoreData.getters.getStoreInformation().isHygieneRequired
+    },
+    hygieneCertificateUntil () {
+      return StoreData.getters.getStoreMember().find(member => member.id === this.user.id)?.hygiene_certificate_until ?? null
+    },
+    isMissingHygieneCertificate () {
+      return this.isHygieneRequired && (
+        !this.hygieneCertificateUntil ||
+        new Date(this.hygieneCertificateUntil) < this.date
+      )
     },
   },
   methods: {