+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Added the option to mark quiz questions as mandatory. Mandatory questions get included every time. #2179 !3760 @AntonBallmaier
 - Added support form for our new zammad ticket system !3210 !3873 @chriswalg @alex.simm
 - Add push notification toggle and prompt users to use push notifications for the chat #1967 !3375 @AntonBallmaier
+- Added store hygiene requirement #1508 !3756 @AntonBallmaier
 
 ## Changes
 - Improve accessibility in create forum thread form !3488 @iron9